+++ apps/openci_runner/lib/src/features/tart_vm/vm_name.dart
@@ -1,3 +1,0 @@
-import 'package:uuid/uuid.dart';
-
-String getVMName() => const Uuid().v4();