+++ apps/openci_runner/lib/src/service/uuid_service.dart
@@ -1,0 +1,3 @@
+import 'package:uuid/uuid.dart';
+
+String get generateUUID => const Uuid().v4();